HIGH: fast sigma clipping implementation added and tested.
I added a full sigma clipping implementation using Frederic's implementation for the standard deviation computations and numba with np.mean and np.partition for the centers computation. Full run is in the order of one second without compilation time and the order of 5 seconds in the first run (so with numba compilation) as opposed to about 1 minute for each iteration in the old code. Created a small test code to assert equality and it seems that they are exaclty equal (still running some other tests to make sure but at least the first ones gave perfect equality. Also, given how the code is written, pretty much any padding choices are available in this new code. Next step is to addapt the code for any kernel shape.