MID: custom kernel usage added. Number of threads improvements. Initial input checks improvements.
You can now use your own custom kernel for the whole sigma clipping operation. Also, can decide on the number of threads used by cv2 and numba. More checks were added to make sure that the input arguments are the correct ones. I now need to change the padding implementation in the standard deviation code as changing the padding right now only does for the mean and median computations. Also, when deciding to use the mean as the centers, I choose use frederic's mean code as it look to be way more efficient (not sure if the thread count for cv2 actually works as that part runs weirdly fast even if I try to set threads to 1).