LOW: changing structure and making sure that code runs properly with custom kernel.
I am changing the structure of the code so that the convolution of frederic can also be use elswhere to calculate the sliding mean. Also checking the reasonning to make sure it works with custom kernels. Still a work in progress as I need to change devices before continuing to code (hence the committing to the server).