HIGH: borders update. Non-uniform kernel fix. Sliding mean update. Thread count update. BORDER CHOICES NEEDS TESTING.
I updated the code so that the border choices are the same for the padding (so median computations) and when using cv2 or scipy.ndimage. Still need to test all the options to see if they do as intended. The sliding mean computation is now always done using Dr. Auchere's convolution code. I also fixed the convolution code for non uniform kernels as I first had to flip the kernel. The thread count choices was also updated, even though I am still not sure if it works as intended when using OpenCV. Most of the docstrings and type anotations were also updated.