MID: n-Dimensional median code. WORK IN PROGRESS.
Started adding a code so that the median can be done for any input array dimensions. Still a work in progress and some of the tests done shows that there is a problem in the code. Not sure where it comes from yet. Only seen it with weighted kernels when the 0. weight is not in the middle of the kernel. Need further testing and understand of the n-dimensional code. Its not really understandable though, at least not yet.