MID: n-dimension sliding median fix. Tests improvements. All tests passed.
I fixed the n-dimensional sliding median computation (had forgotten to flip the kernel). All the tests that I could think of have passed. There are still tests that should be made for higher dimensions with different weights, but I have no clue what to use to do so as I do not know any functions that can compute the sliding median for high dimensions and kernel weights. Seems promissing though.